increase MAX_SLOTS to 65536
will result in about 4MB of usage.

+++ src/thread-manager-128.c
@@ -7,7 +7,7 @@ static atomic_ulong     CACHE_ALIGNED scheduled = 0;
 static uint128_t_padded CACHE_ALIGNED slots[MAX_SLOTS];
 
 void* worker_thread(void* arg) {
-    uint8_t chunk = atomic_inc(&assigned);
+    uint16_t chunk = atomic_inc(&assigned);
     uint128_t thread_id = ((uint128_t)pthread_self()) << 64;
     while (1) {
         uint128_t task = atomic128_load(&slots[chunk]);
@@ -27,7 +27,7 @@ void* worker_thread(void* arg) {
     }
 }
 
-uint8_t thread_pool_schedule_task(TaskFunc func, void* data, TaskDestroy destroy, uint128_t should_exit) {
+uint16_t thread_pool_schedule_task(TaskFunc func, void* data, TaskDestroy destroy, uint128_t should_exit) {
     int64_t data_diff = (data == NULL) ? 0 : (data - (void*)func);
     int64_t destroy_diff = (destroy == NULL) ? 0 : (destroy - func);
     int64_t is_data_diff = llabs(data_diff) < llabs(destroy_diff);
@@ -37,7 +37,7 @@ uint8_t thread_pool_schedule_task(TaskFunc func, void* data, TaskDestroy destroy
     packed |= (uint128_t)(int64_t)(is_data_diff ? destroy : data) >> SCND_PTR_ALIGN_SHIFT << SECOND_PTR;
     packed |= (uint128_t)(int64_t)(is_data_diff ? data_diff : destroy_diff) >> PTR_ALIGN_SHIFT << DIFF_START;
 
-    uint8_t chunk = atomic_inc(&scheduled);
+    uint16_t chunk = atomic_inc(&scheduled);
     uint128_t expected = atomic128_load(&slots[chunk]);
     do {
         if (expected) syscall(SYS_futex, &slots[chunk], FUTEX_WAKE_PRIVATE, 1, NULL, NULL, 0);
@@ -47,7 +47,7 @@ uint8_t thread_pool_schedule_task(TaskFunc func, void* data, TaskDestroy destroy
 }
 
 uint64_t thread_pool_spawn_joinable(TaskFunc func, void* data, TaskDestroy destroy) {
-    uint8_t chunk = thread_pool_schedule_task(func, data, destroy, 1);
+    uint16_t chunk = thread_pool_schedule_task(func, data, destroy, 1);
     thread_pool_new_thread();
     while (!IS_THREAD_ID(atomic128_load(&slots[chunk]))) __builtin_ia32_pause();
     return atomic_exchange((uint64_t*)((char*)&slots[chunk] + 8), EMPTY);
